Move common.go to more sensible places
