+++ internal/scheduler/context/job.go
@@ -9,8 +9,12 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
+	"golang.org/x/exp/maps"
 	v1 "k8s.io/api/core/v1"
 
+	"github.com/armadaproject/armada/internal/common/armadacontext"
+	armadamaps "github.com/armadaproject/armada/internal/common/maps"
+	armadaslices "github.com/armadaproject/armada/internal/common/slices"
 	schedulerconfig "github.com/armadaproject/armada/internal/scheduler/configuration"
 	"github.com/armadaproject/armada/internal/scheduler/interfaces"
 	"github.com/armadaproject/armada/internal/scheduler/internaltypes"
@@ -186,4 +190,62 @@ func JobSchedulingContextFromJob(job *jobdb.Job) *JobSchedulingContext {
 		ResourceRequirements: job.EfficientResourceRequirements(),
 		GangInfo:             gangInfo,
 	}
+}
+
+// PrintJobSummary logs a summary of the job scheduling context
+// It will log a high level summary at Info level, and a list of all queues + jobs affected at debug level
+func PrintJobSummary(ctx *armadacontext.Context, prefix string, jctxs []*JobSchedulingContext) {
+	if len(jctxs) == 0 {
+		return
+	}
+	jobsByQueue := armadaslices.MapAndGroupByFuncs(
+		jctxs,
+		func(jctx *JobSchedulingContext) string {
+			return jctx.Job.Queue()
+		},
+		func(jctx *JobSchedulingContext) *jobdb.Job {
+			return jctx.Job
+		},
+	)
+	resourcesByQueue := armadamaps.MapValues(
+		jobsByQueue,
+		func(jobs []*jobdb.Job) schedulerobjects.ResourceList {
+			rv := schedulerobjects.NewResourceListWithDefaultSize()
+			for _, job := range jobs {
+				rv.AddV1ResourceList(job.ResourceRequirements().Requests)
+			}
+			return rv
+		},
+	)
+	jobCountPerQueue := armadamaps.MapValues(
+		jobsByQueue,
+		func(jobs []*jobdb.Job) int {
+			return len(jobs)
+		},
+	)
+	jobIdsByQueue := armadamaps.MapValues(
+		jobsByQueue,
+		func(jobs []*jobdb.Job) []string {
+			rv := make([]string, len(jobs))
+			for i, job := range jobs {
+				rv[i] = job.Id()
+			}
+			return rv
+		},
+	)
+	summary := fmt.Sprintf(
+		"affected queues %v; resources %v; jobs per queue %v",
+		maps.Keys(jobsByQueue),
+		armadamaps.MapValues(
+			resourcesByQueue,
+			func(rl schedulerobjects.ResourceList) string {
+				return rl.CompactString()
+			},
+		),
+		jobCountPerQueue,
+	)
+	verbose := fmt.Sprintf("affected jobs %v", jobIdsByQueue)
+
+	ctx.Infof("%s %s", prefix, summary)
+	ctx.Debugf("%s %s", prefix, verbose)
 }
